+++ CHANGELOG.md
@@ -71,6 +71,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * [#835](https://github.com/allora-network/allora-chain/pull/835): Introduce the `x/scheduler` module to manage tasks scheduling
 * [#831](https://github.com/allora-network/allora-chain/pull/831) Allow topic creators to update topic configs
 * [#893](https://github.com/allora-network/allora-chain/pull/893) Make `c_norm` editable per topic and add migration v13
+* [#896](https://github.com/allora-network/allora-chain/pull/896) New queries on open submission windows
 
 ### Changed